Add files via upload
The main functions of this script are to subsets the "dhs.RDS" data into the observations within 10 years from each survey, and to create a variable that groups years of birth so that enough number of observed births is available for each birth year group. 

The resulting R data is called "dhs_part10yrs.rds" which is used throughout the main analyses.